requirements.txt
Django>=1.4
South==0.7.3
PIL==1.1.7
django-kombu==0.9.4
django-celery==2.3.3
django-debug-toolbar==0.8.5
git+ssh://git@github.com/MidwestCommunications/django-loci.git@master#egg=loci
simplegeo==3.0.127
geopy==0.94.1
requests==0.11.1
